requirements.txt
asgiref==3.9.1
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.42.2
coverage==7.10.3
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==5.2.5
django-ckeditor==6.7.2
django-cloudinary-storage==0.3.0
django-countries==7.6.1
django-crispy-forms==2.4
django-js-asset==3.1.2
django-widget-tweaks==1.5.0
idna==3.10
iniconfig==2.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
psycopg2==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0
